+++ websocketpp/processors/hybi13.hpp
@@ -630,7 +630,7 @@ protected:
         key.append(constants::handshake_guid);
 
         unsigned char message_digest[20];
-        sha1::calc(key.c_str(),key.length(),message_digest);
+        sha1::calc(key.c_str(),(int)key.length(),message_digest);
         key = base64_encode(message_digest,20);
 
         return lib::error_code();
